requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
uvicorn
fastapi
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.12.0
pulumi-random>=4.18.0
gunicorn==21.2.0
